refactor(pkg): move shift code to pkg
